Refactor: Reuse v6 juno_unsubscribe handler for v7 and v8
