+++ rpc/v7/events.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/juno/jsonrpc"
-	"github.com/NethermindEth/juno/rpc/rpccore"
 )
 
 type Event struct {
@@ -62,27 +61,6 @@ func (h *Handler) SubscribeNewHeads(ctx context.Context) (uint64, *jsonrpc.Error
 		}
 	})
 	return id, nil
-}
-
-func (h *Handler) Unsubscribe(ctx context.Context, id uint64) (bool, *jsonrpc.Error) {
-	w, ok := jsonrpc.ConnFromContext(ctx)
-	if !ok {
-		return false, jsonrpc.Err(jsonrpc.MethodNotFound, nil)
-	}
-	sub, ok := h.subscriptions.Load(id)
-	if !ok {
-		return false, rpccore.ErrInvalidSubscriptionID
-	}
-
-	subs := sub.(*subscription)
-	if !subs.conn.Equal(w) {
-		return false, rpccore.ErrInvalidSubscriptionID
-	}
-
-	subs.cancel()
-	subs.wg.Wait() // Let the subscription finish before responding.
-	h.subscriptions.Delete(id)
-	return true, nil
 }
 
 // unsubscribe assumes h.mu is unlocked. It releases all subscription resources.
